refactor: extract item spawning logic into ItemSpawnService
Created new ItemSpawnService to handle all item generation logic, reducing
DungeonService complexity and improving maintainability.

Changes:
- Created ItemSpawnService with spawnItems(), rollCursedStatus(), and rollEnchantment() methods
- Moved all item template arrays (weapons, armor, potions, scrolls, rings, wands, food, light sources) to ItemSpawnService
- Updated DungeonService to use ItemSpawnService dependency
- Fixed bug in main.ts replaySeed() missing monsterSpawnService parameter
- Removed obsolete cursed-item-generation tests (moved logic to ItemSpawnService)

Benefits:
- DungeonService reduced from 1081 lines to 568 lines (47% reduction)
- Single Responsibility: DungeonService now focuses on structure, ItemSpawnService handles items
- Better separation of concerns and easier to maintain
- All 2187 tests passing
