+++ src/services/ItemSpawnService/index.ts
@@ -1,0 +1,1 @@
+export { ItemSpawnService } from './ItemSpawnService'
